Seal tf.nn's interface.
This moves functions that were defined directly in nn.py
into nn_ops.py.
Change: 139979007
